+++ NEWS.md
@@ -4,9 +4,11 @@
   providers (@fmgithub2017, #3 github).
 * New function `rdb_dimensions()` to request the list of the dimensions of
   the available datasets of the providers (@fmgithub2017, #3 github).
+* New function `rdb_series()` to request the list of the series of
+  the available datasets of the providers (@fmgithub2017, #3 github).
 * Add a clearer error message to the function to_xts() in README (Martin
   Feldkircher, 20200326 email).
-* Update vignette for `rdb_datasets()` and `rdb_dimensions()`.
+* Update vignette for `rdb_datasets()`, `rdb_dimensions()` and `rdb_series()`.
 
 # rdbnomics 0.5.2